Refactor: Consolidate mappers into unified models/mappers structure following DDD principles

+++ src/main/java/com/uptc/queenscorner/models/dtos/requests/AgregarLineaRequest.java
@@ -3,26 +3,26 @@ package com.uptc.queenscorner.models.dtos.requests;
 import java.math.BigDecimal;
 
 public class AgregarLineaRequest {
+    
     private String descripcion;
-    private Integer cantidad;
+    private BigDecimal cantidad;
     private BigDecimal valorUnitario;
-
-    public AgregarLineaRequest() {
-    }
-
-    public AgregarLineaRequest(String descripcion, Integer cantidad, BigDecimal valorUnitario) {
+    
+    public AgregarLineaRequest() {}
+    
+    public AgregarLineaRequest(String descripcion, BigDecimal cantidad, BigDecimal valorUnitario) {
         this.descripcion = descripcion;
         this.cantidad = cantidad;
         this.valorUnitario = valorUnitario;
     }
-
+    
     public String getDescripcion() { return descripcion; }
     public void setDescripcion(String descripcion) { this.descripcion = descripcion; }
-
-    public Integer getCantidad() { return cantidad; }
-    public void setCantidad(Integer cantidad) { this.cantidad = cantidad; }
-
+    
+    public BigDecimal getCantidad() { return cantidad; }
+    public void setCantidad(BigDecimal cantidad) { this.cantidad = cantidad; }
+    
     public BigDecimal getValorUnitario() { return valorUnitario; }
     public void setValorUnitario(BigDecimal valorUnitario) { this.valorUnitario = valorUnitario; }
-}
+}
